Add production KPI fields, kg sales, invoice, financial notes, and house charts
- Production: add populasi, feedQtyKg, goodEggsKg, crackedEggsKg, rejectedEggsKg,
  feedPricePerKg, mortality fields with HD%/Feed Intake/FCR/HPP computed display
- Sales: change unit to kg (Decimal), add invoice number generation and printable invoice sheet
- Dashboard: add Feed Intake and FCR per-house bar charts (HouseBarChart component)
- Stock: egg stock now tracks in kg (goodEggsKg - qtySold kg)
- Financial notes: new OWNER-only Catatan Keuangan page with pemasukan/pengeluaran/saldo
- Middleware + navigation: /financial route protected, added to Lainnya menu

+++ src/app/api/v1/sales/[id]/route.ts
@@ -35,9 +35,8 @@ export async function PUT(
     const record = await updateSale(id, {
       date: body.date,
       customerName: body.customerName,
-      qtySold: body.qtySold !== undefined ? parseInt(body.qtySold) : undefined,
-      unitPrice:
-        body.unitPrice !== undefined ? parseFloat(body.unitPrice) : undefined,
+      qtySold: body.qtySold !== undefined ? parseFloat(body.qtySold) : undefined,
+      unitPrice: body.unitPrice !== undefined ? parseFloat(body.unitPrice) : undefined,
       notes: body.notes,
     });
